// src/api/types/AuthTokenRequestRopc.ts
// This file was auto-generated by Fern from our API Definition.

/**
 * OAuth 2.0 resource owner password credentials grant (ROPC).
 */
export interface AuthTokenRequestRopc {
    /** Client identifier. */
    clientId: string;
    grantType: "password";
    /** Resource owner username. */
    username: string;
    /** Resource owner password. */
    password: string;
    /** Space-separated scopes. Optional. */
    scope?: string;
}